backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
billiard==3.6.4.0
cffi==1.15.1
channels==3.0.5
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.4
mysqlclient==2.1.1
prompt-toolkit==3.0.30
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.0.0
pytz==2022.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
